requirements.txt
Flask==2.2.2
pymongo==4.2.0
python-dotenv==0.20.0
waitress==2.1.2
azure-storage-blob==12.13.1